déplace la sélection de la pièce dans la VuePiece commune
