+++ src/situations/commun/vues/piece.js
@@ -2,7 +2,7 @@ import EventEmitter from 'events';
 
 import 'commun/styles/piece.scss';
 
-import { CHANGEMENT_POSITION } from 'commun/modeles/piece';
+import { CHANGEMENT_POSITION, CHANGEMENT_SELECTION } from 'commun/modeles/piece';
 
 export default class VuePiece extends EventEmitter {
   constructor (piece, depotRessources) {
@@ -34,6 +34,27 @@ export default class VuePiece extends EventEmitter {
 
     this.piece.on(CHANGEMENT_POSITION, (nouvellePosition) => {
       metsAJourPosition(this.$piece, nouvellePosition, dimensionsElementParent);
+    });
+
+    this.$piece.mousedown(e => {
+      this.$piece.stop(true);
+      this.$piece.css('opacity', 1);
+      this.piece.changePosition({
+        x: 100 * parseInt(this.$piece.css('left')) / this.$elementParent.width(),
+        y: 100 * parseInt(this.$piece.css('top')) / this.$elementParent.height()
+      });
+      this.piece.selectionne({
+        x: 100 * e.clientX / this.$elementParent.width(),
+        y: 100 * e.clientY / this.$elementParent.height()
+      });
+    });
+
+    this.$piece.on('dragstart', function (event) { event.preventDefault(); });
+    this.$piece.mouseup(e => { this.piece.deselectionne(); });
+
+    this.piece.on(CHANGEMENT_SELECTION, (selectionnee) => {
+      this.$elementParent.append(this.$piece);
+      this.$piece.toggleClass('selectionnee', selectionnee);
     });
   }
 }
