+++ src/situations/controle/vues/piece.js
@@ -1,7 +1,6 @@
 import 'controle/styles/piece.scss';
 
 import VuePieceCommune from 'commun/vues/piece';
-import { CHANGEMENT_SELECTION } from 'controle/modeles/piece';
 import { DISPARITION_PIECE } from 'controle/modeles/situation';
 
 export function animationInitiale ($element) {
@@ -25,27 +24,6 @@ export default class VuePiece extends VuePieceCommune {
 
   affiche (pointInsertion, $) {
     super.affiche(pointInsertion, $);
-
-    this.$piece.mousedown(e => {
-      this.$piece.stop(true);
-      this.$piece.css('opacity', 1);
-      this.piece.changePosition({
-        x: 100 * parseInt(this.$piece.css('left')) / this.$elementParent.width(),
-        y: 100 * parseInt(this.$piece.css('top')) / this.$elementParent.height()
-      });
-      this.piece.selectionne({
-        x: 100 * e.clientX / this.$elementParent.width(),
-        y: 100 * e.clientY / this.$elementParent.height()
-      });
-    });
-
-    this.$piece.on('dragstart', function (event) { event.preventDefault(); });
-    this.$piece.mouseup(e => { this.piece.deselectionne(); });
-
-    this.piece.on(CHANGEMENT_SELECTION, (selectionnee) => {
-      this.$elementParent.append(this.$piece);
-      this.$piece.toggleClass('selectionnee', selectionnee);
-    });
     this.$piece.show(() => { this.callbackApresApparition(this.$piece); });
 
     this.piece.on(DISPARITION_PIECE, () => {
